Audio Selection and Template saving in JSON
In the control panel, the audio can be changed at any time (before, an explicit local URL had to be given)
Now It is possible to export the project in JSON (without audio, in command line using ExportSaveAsJSON). Then you can open it later on using the control panel. (Note: Images are not saved yet, only their URLs. That means if the image URL changes, it will fail to load it!)